requirements.txt
customtkinter==5.2.2
ImageHash==4.3.2
matplotlib==3.10.3
numpy==2.2.5
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
scipy==1.15.3
tensorflow==2.19.0
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.108
